pyproject.toml
[project]
name = "animrecommendersystem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "fastapi[standard]>=0.120.1",
    "huggingface-hub[hf-xet]>=0.36.0",
    "langchain>=1.0.2",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.1",
    "pandas>=2.3.3",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
]
